fix: keep money as Decimal end to end, and lock rows before mutating balances
The model layer declared Numeric(14, 2), the schemas used Decimal, and
services/finance.py did careful ROUND_HALF_UP arithmetic — then every router
cast the result to float on the way to the database. That round-trip through
binary floating point lost precision on each of the 16 write paths, and it lost
it differently on each backend: Postgres rounds on write where SQLite does not,
so the same sequence of transactions produced different balances depending on
where it ran. That would have surfaced as silent corruption during a
SQLite-to-Postgres migration.

- Money columns are now Mapped[Decimal], not Mapped[float].
- New app/money.py owns to_decimal() and money(); money() quantizes to two
  places so every write to a Numeric(14, 2) column is exact. It sits at the
  package root so routers and services share it without an import cycle.
- All 16 float() casts removed.

Also closes a lost-update race: reads that precede a balance mutation now take
a row lock (get_or_404(..., for_update=True)). Without it two concurrent
withdrawals both passed the sufficiency check and the second write discarded
the first, letting an account be overdrawn. SQLite ignores FOR UPDATE, so this
is a no-op in dev and load-bearing on Postgres.

Verified: 100 credits of Rs.0.07 onto Rs.1000 now land at exactly Rs.1007.00,
where float gives 1007.000000000005. Existing suites stay green (70 API checks,
22 recurring-income checks).

+++ app/routers/recurring_income.py
@@ -16,6 +16,7 @@ from sqlalchemy.orm import Session
 from app.auth import get_current_active_user
 from app.db import get_db
 from app.models import Account, RecurringIncome, User
+from app.routers import money
 from app.schemas import (
     RecurringIncomeCreate,
     RecurringIncomeRead,
@@ -108,7 +109,7 @@ def create_recurring_income(
         user_id=user.id,
         account_id=body.account_id,
         name=body.name,
-        amount=float(body.amount),
+        amount=money(body.amount),
         day_of_month=body.day_of_month,
         category=body.category,
         currency=body.currency,
@@ -142,7 +143,7 @@ def update_recurring_income(
     if "account_id" in data and data["account_id"] is not None:
         _own_account_or_404(db, user, data["account_id"])
     if "amount" in data and data["amount"] is not None:
-        data["amount"] = float(data["amount"])
+        data["amount"] = money(data["amount"])
 
     for field, value in data.items():
         setattr(income, field, value)
